requirements.txt
flask==3.1.1
flask-cors==6.0.1
# flask-socketio=5.5.1
# gevent==25.5.1
google-generativeai==0.8.5
gradio==5.35.0
# gunicorn==23.0.0
langchain-google-vertexai==2.0.27
langchain==0.3.26
langgraph==0.5.1
langsmith==0.4.4
pydantic==2.11.7
python-dotenv==1.1.1
Pillow==11.3.0
opencv-python==4.12.0.88
PyMuPDF==1.26.3
SQLAlchemy==2.0.41
watchdog==6.0.0
waitress==3.0.2
